# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  handler.LoonstrokenImportRequest:
    properties:
      items:
        items:
          additionalProperties: {}
          type: object
        type: array
      userId:
        type: string
    type: object
  handler.LoonstrokenImportResponse:
    properties:
      inserted:
        type: integer
      ok:
        type: boolean
      total:
        type: integer
    type: object
...
  handler.noteCreateBody:
    properties:
      deadline:
        type: string
      inhoud:
        type: string
      kleur:
        type: string
      linkedEventId:
        type: string
      prioriteit:
        type: string
      tags:
        items:
          type: string
        type: array
      titel:
        type: string
    type: object
  handler.noteUpdateBody:
    properties:
      deadline:
        type: string
      inhoud:
        type: string
      isArchived:
        type: boolean
      isPinned:
        type: boolean
      kleur:
        type: string
      linkedEventId:
        type: string
      prioriteit:
        type: string
      tags:
        items:
          type: string
        type: array
      titel:
        type: string
    type: object
...
  model.Loonstrook:
    properties:
      amt_zeerintensief:
        type: number
      bruto_betaling:
        type: number
      bruto_inhouding:
        type: number
      componenten:
        type: object
      cumulatieven:
        type: object
      eju_bedrag:
        type: number
      extra_uren_bedrag:
        type: number
      geimporteerd_op:
        type: string
      id:
        type: string
      jaar:
        type: integer
      loonheffing:
        type: number
      netto:
        type: number
      ort_detail:
        type: object
      ort_totaal:
        type: number
      parttime_factor:
        type: number
      pensioenpremie:
        type: number
      periode:
        type: integer
      periode_label:
        type: string
      reiskosten:
        type: number
      salaris_basis:
        type: number
      schaalnummer:
        type: string
      toeslag_balansvlf:
        type: number
      trede:
        type: string
      type:
        type: string
      user_id:
        type: string
      uurloon:
        type: number
      vakantietoeslag:
        type: number
    type: object
  model.Note:
    properties:
      aangemaakt:
        type: string
      deadline:
        type: string
      gewijzigd:
        type: string
      id:
        type: string
      inhoud:
        type: string
      is_archived:
        type: boolean
      is_pinned:
        type: boolean
      kleur:
        type: string
      linked_event_id:
        type: string
      prioriteit:
        type: string
      tags:
        items:
          type: string
        type: array
      titel:
        type: string
      triage_flag:
        type: boolean
      user_id:
        type: string
    type: object
...
host: localhost:8000
info:
  contact: {}
  description: Backend REST API for Jeffries Homeapp.
  title: Jeffries Homeapp API
  version: "1.0"
paths:
...
  /loonstroken:
    get:
      description: Returns all payslips for the user
      parameters:
      - description: User ID
        in: query
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Loonstrook'
            type: array
        "400":
          description: userId verplicht
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: List loonstroken
      tags:
      - Loonstroken
  /loonstroken/import:
    post:
      consumes:
      - application/json
      description: Bulk upserts payslip data extracted from PDFs
      parameters:
      - description: Import payload containing UserID and Items
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.LoonstrokenImportRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.LoonstrokenImportResponse'
        "400":
          description: Ongeldige JSON of ontbrekende velden
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Import loonstroken
      tags:
      - Loonstroken
  /notes:
    get:
      description: Returns all notes for the user
      parameters:
      - description: User ID
        in: query
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Note'
            type: array
        "400":
          description: userId required
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: List notes
      tags:
      - Notes
    post:
      consumes:
      - application/json
      description: Creates a new note for the user
      parameters:
      - description: User ID
        in: query
        name: userId
        required: true
        type: string
      - description: Note Details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.noteCreateBody'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Note'
        "400":
          description: userId required or invalid JSON
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Create a note
      tags:
      - Notes
  /notes/{id}:
    delete:
      description: Permanently deletes a note by its ID
      parameters:
      - description: Note ID (UUID)
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: status deleted
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: invalid id
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Delete a note
      tags:
      - Notes
    get:
      description: Returns a single note by its ID
      parameters:
      - description: Note ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Note'
        "400":
          description: invalid id
          schema:
            type: string
        "404":
          description: note not found
          schema:
            type: string
      summary: Get a note
      tags:
      - Notes
    patch:
      consumes:
      - application/json
      description: Updates the details of an existing note
      parameters:
      - description: Note ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Updated Note Fields
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.noteUpdateBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Note'
        "400":
          description: invalid JSON or ID
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Update a note
      tags:
      - Notes
  /notes/{id}/backlinks:
    get:
      description: Returns notes that reference the given note
      parameters:
      - description: Note ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Note'
            type: array
        "400":
          description: invalid id
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get note backlinks
      tags:
      - Notes
  /notes/search:
    get:
      description: Performs a full-text search across all notes for a user
      parameters:
      - description: User ID
        in: query
        name: userId
        required: true
        type: string
      - description: Search query
        in: query
        name: q
        required: true
        type: string
      - default: 20
        description: Limit count
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Note'
            type: array
        "400":
          description: userId and q required
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Search notes
      tags:
      - Notes
  /notes/tags:
    get:
      description: Returns all unique tags used across a user's notes
      parameters:
      - description: User ID
        in: query
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
        "400":
          description: userId required
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get all note tags
      tags:
      - Notes
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: X-API-Key
    type: apiKey
swagger: "2.0"